3113bf1758d233e35d6547c831e7083e4b003ff4 fix(types): use f64 for DecodePsbt fee field (deadmanoz)

Pull request description:

  As per raised in https://github.com/rust-bitcoin/corepc/issues/429, addressing type discrepancies between `corepc-types` and Bitcoin Core

  Bitcoin Core returns the `fee` field in the `decodepsbt` RPC response as a floating-point (double) BTC value, not as satoshis (integer). The current implementation uses `u64` for the fee field and converts it with `Amount::from_sat()`, which causes a deserialisation failure.

  Minimal repro with a PSBT from Bitcoin Core's functional tests.

  https://github.com/bitcoin/bitcoin/blob/13891a8a685d255cb13dd5018e3d5ccc18b07c34/test/functional/data/rpc_psbt.json#L96

  ```rust
  use corepc_client::client_sync::v28::Client;
  use corepc_client::client_sync::Auth;
  use corepc_client::types::v28::DecodePsbt;

  fn main() -> Result<(), Box<dyn std::error::Error>> {
      // First valid PSBT from Bitcoin Core's test/functional/data/rpc_psbt.json (valid[0]).
      let psbt = concat!(
          "cHNidP8BAHUCAAAAASaBcTce3/KF6Tet7qSze3gADAVmy7OtZGQXE8pCFxv2AAAAAAD+////",
          "AtPf9QUAAAAAGXapFNDFmQPFusKGh2DpD9UhpGZap2UgiKwA4fUFAAAAABepFDVF5uM7gyxH",
          "BQ8k0+65PJwDlIvHh7MuEwAAAQD9pQEBAAAAAAECiaPHHqtNIOA3G7ukzGmPopXJRjr6Ljl/",
          "hTPMti+VZ+UBAAAAFxYAFL4Y0VKpsBIDna89p95PUzSe7LmF/////4b4qkOnHf8USIk6Uwpy",
          "N+9rRgi7st0tAXHmOuxqSJC0AQAAABcWABT+Pp7xp0XpdNkCxDVZQ6vLNL1TU/////8CAMLr",
          "CwAAAAAZdqkUhc/xCX/Z4Ai7NK9wnGIZeziXikiIrHL++E4sAAAAF6kUM5cluiHv1irHU6m8",
          "0GfWx6ajnQWHAkcwRAIgJxK+IuAnDzlPVoMR3HyppolwuAJf3TskAinwf4pfOiQCIAGLONfc",
          "0xTnNMkna9b7QPZzMlvEuqFEyADS8vAtsnZcASED0uFWdJQbrUqZY3LLh+GFbTZSYG2YVi/j",
          "nF6efkE/IQUCSDBFAiEA0SuFLYXc2WHS9fSrZgZU327tzHlMDDPOXMMJ/7X85Y0CIGczio4O",
          "FyXBl/saiK9Z9R5E5CVbIBZ8hoQDHAXR8lkqASECI7cr7vCWXRC+B3jv7NYfysb3mk6haTkz",
          "gHNEZPhPKrMAAAAAAAAA",
      );

      let auth = Auth::UserPass("bitcoin".to_string(), "bitcoin".to_string());
      let client = Client::new_with_auth("http://127.0.0.1:8332", auth)?;
      let decoded: DecodePsbt = client.call("decodepsbt", &[serde_json::json!(psbt)])?;
      println!("fee: {:?}", decoded.fee);
      Ok(())
  }
  ```

  ```
  Error: JsonRpc(Json(Error("invalid type: floating point `3.01e-6`, expected u64", line: 1, column: 2896)))
  ```

  This change:
  - Changes the fee field type from `u64` to `f64` in `DecodePsbt` struct (as required in v17, v23, v24, v30)
  - Updates the conversion to use `Amount::from_btc()` instead of `Amount::from_sat()`
  - Adds a `Fee(ParseAmountError)` error variant to `DecodePsbtError` for proper error handling of the fee
  - Adds an integration test to verify correct deserialisation

ACKs for top commit:
  tcharding:
    ACK 3113bf1758d233e35d6547c831e7083e4b003ff4

Tree-SHA512: de6ec1c23869df79bfb7317831531d42a1c08602af77ff1575a514f47356112f55f5327a9d6fa7923fddb5e3253768ae5013454439ff29a4fe1aa4075df1c66c

+++ types/src/v23/raw_transactions/mod.rs
@@ -47,7 +47,7 @@ pub struct DecodePsbt {
     /// Array of transaction outputs.
     pub outputs: Vec<PsbtOutput>,
     /// The transaction fee paid if all UTXOs slots in the PSBT have been filled.
-    pub fee: Option<u64>,
+    pub fee: Option<f64>,
 }
 
 /// An item from the global xpubs list. Part of `decodepsbt`.
